There is the main gui and the settings gui
