+++ src/input/gamepad.rs
@@ -10,13 +10,7 @@ pub trait GamepadState {
 }
 
 pub trait Gamepads {
-    type State;
-    fn advance(
-        &mut self,
-        gamepad_event: &GamepadEvent,
-        input_settings: &mut InputSettings,
-        state: &mut Self::State,
-    );
+    fn advance(&mut self, gamepad_event: &GamepadEvent, input_settings: &mut InputSettings);
     fn get_joypad(&mut self, id: &InputId, mapping: &JoypadGamepadMapping) -> JoypadState;
     fn get_gamepad_by_input_id(&self, id: &InputId) -> Option<&dyn GamepadState>;
 }
